requirements.txt
python-dotenv>=0.19.0
vkbottle>=4.0.0,<5.0.0
requests>=2.25.0

# Для обработки документов
PyPDF2>=3.0.0
python-docx>=0.8.11

# Для OCR (опционально - если нужно распознавание текста с изображений)
# Pillow>=9.0.0
# pytesseract>=0.3.10

# Deepseek API использует requests, который уже установлен